Add migration-tool for aufs to overlay2
Implements storage-driver migration without the need to pull layer data again.
To achieve this we translate the on-disk AUFS file tree to the format used by
overlayFS. This includes containers that exist at the time of the migration.

To make these changes effective the engine has to be reconfigured to use
the overlay2 storage-driver and containers have to be restarted.

Signed-off-by: Robert Günzler <robertg@balena.io>